bbc/requirements/requirements.txt
baluhn==0.1
beautifulsoup4==4.4.1
Django==1.8.5
django-bootstrap-form==3.2
django-localflavor==1.2
functools32==3.2.3.post2
jdt==0.0.3
jsonschema==2.5.1
oauthlib==1.0.3
pdt==0.7.4
PyJWT==1.4.0
pymongo==3.2.2
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.9.1
requests-oauthlib==0.8.0
six==1.10.0
validate-email==1.3
wheel==0.24.0
